+++ pkg/coderefs/coderefs_test.go
@@ -1,6 +1,7 @@
 package coderefs
 
 import (
+	"os"
 	"regexp"
 	"sort"
 	"strings"
@@ -27,92 +28,120 @@ func init() {
 	log.Init(true)
 }
 
-func Test_generateReferencesFromGrep(t *testing.T) {
+func delimit(s string, delim string) string {
+	return delim + s + delim
+}
+
+const (
+	testFlagKey  = "someFlag"
+	testFlagKey2 = "anotherFlag"
+)
+
+func TestMain(m *testing.M) {
+	log.Init(true)
+	os.Exit(m.Run())
+}
+
+func Test_generateReferences(t *testing.T) {
+	testResult := []string{"", "flags.txt", ":", "12", delimit(testFlagKey, `"`)}
+	testWant := searchResultLine{Path: "flags.txt", LineNum: 12, LineText: delimit(testFlagKey, `"`), FlagKeys: []string{testFlagKey}}
+
 	tests := []struct {
-		name       string
-		flags      []string
-		grepResult [][]string
-		ctxLines   int
-		want       []grepResultLine
-		exclude    string
+		name         string
+		flags        []string
+		searchResult [][]string
+		ctxLines     int
+		want         []searchResultLine
+		exclude      string
 	}{
 		{
-			name:  "succeeds",
-			flags: []string{"someFlag", "anotherFlag"},
-			grepResult: [][]string{
-				{"", "flags.txt", ":", "12", "someFlag"},
-			},
-			ctxLines: 0,
-			want: []grepResultLine{
-				{Path: "flags.txt", LineNum: 12, LineText: "someFlag", FlagKeys: []string{"someFlag"}},
-			},
+			name:         "succeeds",
+			flags:        []string{testFlagKey, testFlagKey2},
+			searchResult: [][]string{testResult},
+			ctxLines:     0,
+			want:         []searchResultLine{testWant},
 		},
 		{
-			name:  "succeeds with exclude",
-			flags: []string{"someFlag", "anotherFlag"},
-			grepResult: [][]string{
-				{"", "flags.txt", ":", "12", "someFlag"},
-			},
-			ctxLines: 0,
-			want:     []grepResultLine{},
-			exclude:  ".*",
+			name:         "succeeds with exclude",
+			flags:        []string{testFlagKey, testFlagKey2},
+			searchResult: [][]string{testResult},
+			ctxLines:     0,
+			want:         []searchResultLine{},
+			exclude:      ".*",
 		},
 		{
-			name:  "succeeds with no LineText lines",
-			flags: []string{"someFlag", "anotherFlag"},
-			grepResult: [][]string{
-				{"", "flags.txt", ":", "12", "someFlag"},
-			},
-			ctxLines: -1,
-			want: []grepResultLine{
-				{Path: "flags.txt", LineNum: 12, FlagKeys: []string{"someFlag"}},
+			name:         "succeeds with no LineText lines",
+			flags:        []string{testFlagKey, testFlagKey2},
+			searchResult: [][]string{testResult},
+			ctxLines:     -1,
+			want: []searchResultLine{
+				{Path: "flags.txt", LineNum: 12, FlagKeys: []string{testFlagKey}},
 			},
 		},
 		{
 			name:  "succeeds with multiple references",
-			flags: []string{"someFlag", "anotherFlag"},
-			grepResult: [][]string{
-				{"", "flags.txt", ":", "12", "someFlag"},
-				{"", "path/flags.txt", ":", "12", "someFlag anotherFlag"},
+			flags: []string{testFlagKey, testFlagKey2},
+			searchResult: [][]string{
+				testResult,
+				{"", "path/flags.txt", ":", "12", `"someFlag" "anotherFlag"`},
 			},
 			ctxLines: 0,
-			want: []grepResultLine{
-				{Path: "flags.txt", LineNum: 12, LineText: "someFlag", FlagKeys: []string{"someFlag"}},
-				{Path: "path/flags.txt", LineNum: 12, LineText: "someFlag anotherFlag", FlagKeys: []string{"someFlag", "anotherFlag"}},
+			want: []searchResultLine{
+				testWant,
+				{Path: "path/flags.txt", LineNum: 12, LineText: `"someFlag" "anotherFlag"`, FlagKeys: []string{testFlagKey, testFlagKey2}},
 			},
 		},
 		{
 			name:  "succeeds with extra LineText lines",
-			flags: []string{"someFlag", "anotherFlag"},
-			grepResult: [][]string{
+			flags: []string{testFlagKey, testFlagKey2},
+			searchResult: [][]string{
 				{"", "flags.txt", "-", "11", "not a flag key line"},
-				{"", "flags.txt", ":", "12", "someFlag"},
+				testResult,
 				{"", "flags.txt", "-", "13", "not a flag key line"},
 			},
 			ctxLines: 1,
-			want: []grepResultLine{
+			want: []searchResultLine{
 				{Path: "flags.txt", LineNum: 11, LineText: "not a flag key line"},
-				{Path: "flags.txt", LineNum: 12, LineText: "someFlag", FlagKeys: []string{"someFlag"}},
+				testWant,
 				{Path: "flags.txt", LineNum: 13, LineText: "not a flag key line"},
 			},
 		},
 		{
 			name:  "succeeds with extra LineText lines and multiple flags",
-			flags: []string{"someFlag", "anotherFlag"},
-			grepResult: [][]string{
+			flags: []string{testFlagKey, testFlagKey2},
+			searchResult: [][]string{
 				{"", "flags.txt", "-", "11", "not a flag key line"},
-				{"", "flags.txt", ":", "12", "someFlag"},
+				testResult,
 				{"", "flags.txt", "-", "13", "not a flag key line"},
-				{"", "flags.txt", ":", "14", "anotherFlag"},
+				{"", "flags.txt", ":", "14", delimit(testFlagKey2, `"`)},
 				{"", "flags.txt", "-", "15", "not a flag key line"},
 			},
 			ctxLines: 1,
-			want: []grepResultLine{
+			want: []searchResultLine{
 				{Path: "flags.txt", LineNum: 11, LineText: "not a flag key line"},
-				{Path: "flags.txt", LineNum: 12, LineText: "someFlag", FlagKeys: []string{"someFlag"}},
+				testWant,
 				{Path: "flags.txt", LineNum: 13, LineText: "not a flag key line"},
-				{Path: "flags.txt", LineNum: 14, LineText: "anotherFlag", FlagKeys: []string{"anotherFlag"}},
+				{Path: "flags.txt", LineNum: 14, LineText: delimit(testFlagKey2, `"`), FlagKeys: []string{testFlagKey2}},
 				{Path: "flags.txt", LineNum: 15, LineText: "not a flag key line"},
+			},
+		},
+		{
+			name:         "does not match substring flag key",
+			flags:        []string{testFlagKey, testFlagKey[:4]},
+			searchResult: [][]string{testResult},
+			ctxLines:     0,
+			want:         []searchResultLine{testWant},
+		},
+		{
+			// delimeters don't have to match on both sides
+			name:  "succeeds with multiple delimiters",
+			flags: []string{"someFlag", "some"},
+			searchResult: [][]string{
+				{"", "flags.txt", ":", "12", `"` + testFlagKey + "'"},
+			},
+			ctxLines: 0,
+			want: []searchResultLine{
+				{Path: "flags.txt", LineNum: 12, LineText: `"` + testFlagKey + "'", FlagKeys: []string{testFlagKey}},
 			},
 		},
 	}
@@ -120,7 +149,7 @@ func Test_generateReferencesFromGrep(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			ex, err := regexp.Compile(tt.exclude)
 			require.NoError(t, err)
-			got := generateReferencesFromGrep(tt.flags, tt.grepResult, tt.ctxLines, ex)
+			got := generateReferences(tt.flags, tt.searchResult, tt.ctxLines, `"'`, ex)
 			require.Equal(t, tt.want, got)
 		})
 	}
@@ -134,12 +163,12 @@ func Test_findReferencedFlags(t *testing.T) {
 	}{
 		{
 			name: "finds a flag",
-			ref:  "line contains someFlag",
+			ref:  "line contains " + delimit(testFlagKey, `"`),
 			want: []string{"someFlag"},
 		},
 		{
 			name: "finds multiple flags",
-			ref:  "line contains someFlag and anotherFlag",
+			ref:  "line contains " + delimit(testFlagKey, `"`) + " " + delimit(testFlagKey2, `"`),
 			want: []string{"someFlag", "anotherFlag"},
 		},
 		{
@@ -150,7 +179,7 @@ func Test_findReferencedFlags(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := findReferencedFlags(tt.ref, []string{"someFlag", "anotherFlag"})
+			got := findReferencedFlags(tt.ref, []string{testFlagKey, testFlagKey2}, `"`)
 			require.Equal(t, tt.want, got)
 		})
 	}
@@ -161,30 +190,30 @@ func Test_makeReferenceHunksReps(t *testing.T) {
 
 	tests := []struct {
 		name string
-		refs grepResultLines
+		refs searchResultLines
 		want []ld.ReferenceHunksRep
 	}{
 		{
 			name: "no references",
-			refs: grepResultLines{},
+			refs: searchResultLines{},
 			want: []ld.ReferenceHunksRep{},
 		},
 		{
 			name: "single path, single reference with context lines",
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "context +1",
@@ -207,14 +236,14 @@ func Test_makeReferenceHunksReps(t *testing.T) {
 		},
 		{
 			name: "multiple paths, single reference with context lines",
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  1,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/c/d",
 					LineNum:  10,
 					LineText: "flag-2",
@@ -263,26 +292,26 @@ func Test_makeHunkReps(t *testing.T) {
 	tests := []struct {
 		name     string
 		ctxLines int
-		refs     grepResultLines
+		refs     searchResultLines
 		want     []ld.HunkRep
 	}{
 		{
 			name:     "single reference with context lines",
 			ctxLines: 1,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "context +1",
@@ -301,32 +330,32 @@ func Test_makeHunkReps(t *testing.T) {
 		{
 			name:     "multiple references, single flag, one hunk",
 			ctxLines: 1,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "context inner",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  8,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  9,
 					LineText: "context +1",
@@ -345,32 +374,32 @@ func Test_makeHunkReps(t *testing.T) {
 		{
 			name:     "multiple references, multiple context lines, single flag, one hunk",
 			ctxLines: 2,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "context inner",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  8,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  9,
 					LineText: "context +1",
@@ -389,38 +418,38 @@ func Test_makeHunkReps(t *testing.T) {
 		{
 			name:     "multiple references, single flag, multiple hunks",
 			ctxLines: 1,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "a context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "a flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "a context +1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  9,
 					LineText: "b context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  10,
 					LineText: "b flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  11,
 					LineText: "b context +1",
@@ -445,32 +474,32 @@ func Test_makeHunkReps(t *testing.T) {
 		{
 			name:     "multiple consecutive references, multiple flags, multiple hunks",
 			ctxLines: 1,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "context inner",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  8,
 					LineText: "flag-2",
 					FlagKeys: []string{"flag-2"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  9,
 					LineText: "context +1",
@@ -495,38 +524,38 @@ func Test_makeHunkReps(t *testing.T) {
 		{
 			name:     "multiple consecutive (non overlapping) references, multiple flags, multiple hunks",
 			ctxLines: 1,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "a context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "a flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "a context +1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  8,
 					LineText: "b context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  9,
 					LineText: "b flag-2",
 					FlagKeys: []string{"flag-2"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  10,
 					LineText: "b context +1",
@@ -551,32 +580,32 @@ func Test_makeHunkReps(t *testing.T) {
 		{
 			name:     "multiple references, single flag, 0 context, multiple hunks",
 			ctxLines: 0,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "context inner",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  8,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  9,
 					LineText: "context +1",
@@ -601,32 +630,32 @@ func Test_makeHunkReps(t *testing.T) {
 		{
 			name:     "multiple references, single flag, negative context, multiple hunks",
 			ctxLines: -1,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  5,
 					LineText: "context -1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  6,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  7,
 					LineText: "context inner",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  8,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  9,
 					LineText: "context +1",
@@ -656,38 +685,38 @@ func Test_makeHunkReps(t *testing.T) {
 			// times forwards, which would walk past the correct end of the hunk and into the next hunk
 			name:     "multiple references, first reference at start of file",
 			ctxLines: 1,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  1,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  2,
 					LineText: "context+1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  3,
 					LineText: "context+3",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  10,
 					LineText: "context-1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  11,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  12,
 					LineText: "context+1",
@@ -714,32 +743,32 @@ func Test_makeHunkReps(t *testing.T) {
 			// previous test case
 			name:     "multiple references, first reference at start of file",
 			ctxLines: 2,
-			refs: grepResultLines{
-				grepResultLine{
+			refs: searchResultLines{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  1,
 					LineText: "context-1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  2,
 					LineText: "flag-1",
 					FlagKeys: []string{"flag-1"},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  3,
 					LineText: "context+1",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  4,
 					LineText: "context+2",
 					FlagKeys: []string{},
 				},
-				grepResultLine{
+				searchResultLine{
 					Path:     "a/b",
 					LineNum:  10,
 					LineText: "context+alot+shouldn'tbeinhunk",
@@ -763,9 +792,9 @@ func Test_makeHunkReps(t *testing.T) {
 
 			require.Equal(t, len(groupedResults), 1)
 
-			fileGrepResults := groupedResults[0]
+			fileSearchResults := groupedResults[0]
 
-			got := fileGrepResults.makeHunkReps(projKey, tt.ctxLines)
+			got := fileSearchResults.makeHunkReps(projKey, tt.ctxLines)
 
 			sort.Sort(byStartingLineNumber(got))
 
@@ -775,38 +804,38 @@ func Test_makeHunkReps(t *testing.T) {
 }
 
 func Test_groupIntoPathMap(t *testing.T) {
-	grepResultPathALine1 := grepResultLine{
+	searchResultPathALine1 := searchResultLine{
 		Path:     "a",
 		LineNum:  1,
 		LineText: "flag-1",
 		FlagKeys: []string{"flag-1"},
 	}
 
-	grepResultPathALine2 := grepResultLine{
+	searchResultPathALine2 := searchResultLine{
 		Path:     "a",
 		LineNum:  2,
 		LineText: "flag-2",
 		FlagKeys: []string{"flag-2"},
 	}
 
-	grepResultPathBLine1 := grepResultLine{
+	searchResultPathBLine1 := searchResultLine{
 		Path:     "b",
 		LineNum:  1,
 		LineText: "flag-3",
 		FlagKeys: []string{"flag-3"},
 	}
-	grepResultPathBLine2 := grepResultLine{
+	searchResultPathBLine2 := searchResultLine{
 		Path:     "b",
 		LineNum:  2,
 		LineText: "flag-2",
 		FlagKeys: []string{"flag-4"},
 	}
 
-	lines := grepResultLines{
-		grepResultPathALine1,
-		grepResultPathALine2,
-		grepResultPathBLine1,
-		grepResultPathBLine2,
+	lines := searchResultLines{
+		searchResultPathALine1,
+		searchResultPathALine2,
+		searchResultPathBLine1,
+		searchResultPathBLine2,
 	}
 
 	linesByPath := lines.aggregateByPath()
@@ -820,10 +849,10 @@ func Test_groupIntoPathMap(t *testing.T) {
 	require.Contains(t, aRefMap, "flag-1")
 	require.Contains(t, aRefMap, "flag-2")
 
-	aLines := aRefs.fileGrepResultLines
+	aLines := aRefs.fileSearchResultLines
 	require.Equal(t, aLines.Len(), 2)
-	require.Equal(t, aLines.Front().Value, grepResultPathALine1)
-	require.Equal(t, aLines.Back().Value, grepResultPathALine2)
+	require.Equal(t, aLines.Front().Value, searchResultPathALine1)
+	require.Equal(t, aLines.Back().Value, searchResultPathALine2)
 
 	bRefs := linesByPath[1]
 	require.Equal(t, bRefs.path, "b")
@@ -834,10 +863,10 @@ func Test_groupIntoPathMap(t *testing.T) {
 	require.Contains(t, bRefMap, "flag-3")
 	require.Contains(t, bRefMap, "flag-4")
 
-	bLines := bRefs.fileGrepResultLines
+	bLines := bRefs.fileSearchResultLines
 	require.Equal(t, bLines.Len(), 2)
-	require.Equal(t, bLines.Front().Value, grepResultPathBLine1)
-	require.Equal(t, bLines.Back().Value, grepResultPathBLine2)
+	require.Equal(t, bLines.Front().Value, searchResultPathBLine1)
+	require.Equal(t, bLines.Back().Value, searchResultPathBLine2)
 }
 
 func Test_filterShortFlags(t *testing.T) {